# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HumanTouch API",
    "description": "AI内容人性化处理系统 - 将AI生成的文本转换为更自然的人类写作风格，有效规避AI检测工具。",
    "version": "1.0.0",
    "contact": {
      "name": "HumanTouch Support",
      "email": "support@humantouch.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.humantouch.dev",
      "description": "生产环境"
    },
    {
      "url": "http://localhost:3000",
      "description": "本地开发环境"
    }
  ],
  "paths": {
...
    "/api/v1/status/{taskId}": {
      "get": {
        "tags": ["Async Processing"],
        "summary": "查询任务状态",
        "description": "查询异步任务的处理状态和结果。",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "任务ID"
          }
        ],
        "responses": {
          "200": {
            "description": "任务状态",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/tasks": {
      "get": {
        "tags": ["Async Processing"],
        "summary": "获取任务列表",
        "description": "获取用户的任务列表，支持分页和状态过滤。",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["pending", "processing", "completed", "failed"]
            },
            "description": "任务状态过滤"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            },
            "description": "返回结果数量限制"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "分页偏移量"
          }
        ],
        "responses": {
          "200": {
            "description": "任务列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/validate": {
      "post": {
        "tags": ["Detection"],
        "summary": "AI检测验证",
        "description": "使用多个AI检测工具验证文本的AI检测分数。",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "检测结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ProcessResponse": {
        "type": "object",
        "properties": {
          "processed_text": {
            "type": "string",
            "description": "处理后的文本"
          },
          "original_length": {
            "type": "integer",
            "description": "原始文本长度"
          },
          "processed_length": {
            "type": "integer",
            "description": "处理后文本长度"
          },
          "detection_scores": {
            "type": "object",
            "properties": {
              "zerogpt": {
                "type": "number",
                "description": "ZeroGPT检测分数"
              },
              "gptzero": {
                "type": "number",
                "description": "GPTZero检测分数"
              },
              "copyleaks": {
                "type": "number",
                "description": "Copyleaks检测分数"
              }
            }
          },
          "processing_time": {
            "type": "number",
            "description": "处理耗时（秒）"
          },
          "rounds_used": {
            "type": "integer",
            "description": "实际使用的处理轮数"
          }
        }
      },
...
      "TaskStatusResponse": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/ProcessResponse"
          },
          "error": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TaskListResponse": {
        "type": "object",
        "properties": {
          "tasks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "task_id": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "text_preview": {
                  "type": "string"
                }
              }
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "offset": {
                "type": "integer"
              },
              "has_more": {
                "type": "boolean"
              }
            }
          }
        }
      },
      "ValidateRequest": {
        "type": "object",
        "required": ["text"],
        "properties": {
          "text": {
            "type": "string",
            "minLength": 10,
            "maxLength": 10000
          },
          "detectors": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["zerogpt", "gptzero", "copyleaks"]
            },
            "description": "使用的检测工具"
          }
        }
      },
      "ValidateResponse": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "detection_scores": {
            "type": "object",
            "additionalProperties": {
              "type": "number"
            }
          },
          "summary": {
            "type": "object",
            "properties": {
              "overall_score": {
                "type": "number"
              },
              "human_likelihood": {
                "type": "number"
              }
            }
          }
        }
      },
...
    },
    "responses": {
...
      "NotFound": {
        "description": "资源未找到",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}